feat: Electra (#45)
* Bump golibp2p

* refactor: Remove unused Prater chain configuration

* refactor: Remove unused Prater network configuration

* feat: Add Go setup and release workflows

* limit attestations and sync committees to 2 subnets

* feat: add Electra fork version support

This commit introduces support for the Electra fork version in
hermes. It includes updates to handle the new fork version in
genesis, pubsub, and command-line tools. The go.mod file is also
updated to include the latest dependencies.

build: update go.sum dependencies

This commit updates the dependencies in go.sum to their latest
versions. This ensures that the project is using the most up-to-date
versions of its dependencies, which can improve security,
performance, and stability.

build: update go.sum dependencies to latest versions

This commit updates the dependencies in go.sum to their latest
versions. This ensures that the project is using the most up-to-date
and secure versions of its dependencies.

build: update k8s dependencies in go.sum to v0.30.4

This commit updates the Kubernetes dependencies in the go.sum file
to version v0.30.4. This ensures that the project is using the
latest versions of the Kubernetes libraries.

* feat: add drips funding info and remove retro funding info

* feat: add support for Optimism RetroPGF-3 funding round

* fix(host): remove git merge conflict markers in host.go

* feat: add Electra fork version support

This commit introduces support for the Electra fork version throughout the codebase. It includes updates to genesis handling, Kinesis output rendering, and request/response mechanisms to correctly process and manage Electra blocks.

* chore: remove funding.json and FUNDING.json files

These files are no longer needed.

* feat(eth): adds handlers for new electra attestations

* feat: Add support for rendering SingleAttestation

* feat(kinesis): support electra signed block

* revert(config): limit attestations and sync committees to 2 subnets

This reverts commit ad094487188b9bc507eea3e44feb7b351d7c37d3.

* revert(build): feat: Add Go setup and release workflows

This reverts commit 1685c1c2eeca29e1dc3759703993d2c854c36651.

* chore(hermes): updates go-ethereum

* fix: Update required Go version to 1.23

* fix: Update ethereum/go-ethereum to v1.15.5

* add funding

* add funding

---------

Co-authored-by: Matty Evans <mattevansnz@gmail.com>
Co-authored-by: Mikel Cortes <45786396+cortze@users.noreply.github.com>