feat: add configurable subnets for gossip topics
This commit introduces the ability to configure which subnets a node
subscribes to for gossip topics. This allows for more granular control
over network traffic and resource usage.

The following changes were made:

- Added a `SubnetConfig` struct to represent the configuration for a
  subnet.
- Added a `SubnetSelectionType` enum to represent the different types
  of subnet selection.
- Added a `SubnetConfigs` field to the `NodeConfig` struct to store the
  subnet configurations.
- Added a `HasSubnets` function to determine if a gossip topic has
  subnets, and if so, how many.
- Added a `Validate` method to the `SubnetConfig` struct to validate
  the configuration.
- Added a `GetSubscribedSubnets` function to return the subnet IDs to
  subscribe to for a given topic.
- Modified the `getDesiredFullTopics` method to use the subnet
  configurations to determine which subnets to subscribe to.
- Added unit tests for the new functionality.

This change allows operators to reduce the number of gossip topics a
node subscribes to, which can reduce network traffic and resource usage.
For example, a node that is only interested in a subset of attestations
can configure itself to only subscribe to the subnets that are relevant
to it.