feat(eth): add support for subscribing to custom topics (#62)
* feat(eth): add support for subscribing to custom topics

This commit adds a new flag `--subscription.topics` to the `eth` command, which allows users to specify a comma-separated list of topics to subscribe to. If this flag is not set, the default list of topics will be used.

This change is useful for users who only want to subscribe to a subset of topics, which can reduce the amount of data they receive and improve performance.

* docs: add documentation for topic subscription feature
refactor(node_config): simplify topic subscription check to improve readability